Fix some x64 specific C4244 warnings. Patch by XhmikosR.
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@4699 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/apps/mplayerc/LineNumberEdit.cpp
@@ -394,7 +394,7 @@ LRESULT CLineNumberEdit::OnSelectLine(WPARAM wParam, LPARAM /*lParam*/ )
 {
 
 	// Calc start and end position of the line
-	int lineno = wParam + GetScrollPos( SB_VERT );
+	int lineno = static_cast<int>(wParam) + GetScrollPos( SB_VERT );
 	int start = LineIndex( lineno );
 	int end = LineIndex( lineno + 1 );
 	SetSel( start, end - 1 );
